+++ CHANGELOG.md
@@ -11,7 +11,9 @@ Subheadings to categorize changes are `added, changed, deprecated, removed, fixe
 - Added `data-loader-shim` to workers to create shim script.
 - Added tailwindcss support via `rel="tailwind-css"`.
 - Added support for `svg` files when using `rel="inline"`
-- Print all acessible addresses if `0.0.0.0` is used.
+- Print all accessible addresses if `0.0.0.0` is used.
+- Added HTML & CSS minification when serving `--release` mode.
+- Added PNG optimization when serving in `--release` mode.
 
 ### changed
 - Updated gloo-worker example to use gloo-worker crate v2.1.